Add: 删除无用SSD1306程序

+++ u8g2/u8x8_d_ssd1306_128x64_noname.c
@@ -264,7 +264,7 @@ static uint8_t u8x8_d_ssd1306_sh1106_generic(u8x8_t *u8x8, uint8_t msg, uint8_t 
       {
 	c = ((u8x8_tile_t *)arg_ptr)->cnt;
 	ptr = ((u8x8_tile_t *)arg_ptr)->tile_ptr;
-	u8x8_cad_SendData(u8x8, c*8, ptr); 	/* note: SendData can not handle more than 255 bytes */
+	u8x8_cad_SendData(u8x8, (uint8_t)(c*8), ptr); 	/* note: SendData can not handle more than 255 bytes */
 	/*
 	do
 	{
